Projeto_biblioteca/requirements.txt
sqlalchemy==2.0.35
pandas==2.2.2
python-dotenv==1.0.1
pyodbc==5.1.0